Implement route user/id

+++ src/main/java/com/chatop/portal/repository/RentalsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.chatop.portal.entity.Rentals;
 
 @Repository // Declare this interface as a repository
-public interface RentalsRepository extends JpaRepository<Rentals, Long> {
+public interface RentalsRepository extends JpaRepository<Rentals, Integer> {
 
 }
